requirements.txt
django
django-admin
django-allauth
djangorestframework
pydotenv
python-dotenv
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
fpdf
